app/requirements.txt
Flask==3.0.3
py-healthcheck==1.10.1
prometheus-client==0.20.0
#psycopg2==2.9.9 # Package for database